+++ src/Renderer.h
@@ -4,4 +4,4 @@
 #include <Eigen.h>
 #include "Face.h"
 
-std::tuple<cv::Mat, cv::Mat, cv::Mat, cv::Mat> render(Face& face, Matrix4f projectionMatrix, int height, int width);
+std::tuple<cv::Mat, cv::Mat, cv::Mat, cv::Mat> render(Face& face, Matrix4d projectionMatrix, int height, int width);
